Move APIs needed by Rive to public
Move GetScaleFactor to public
