+++ engine/graphics/src/dmsdk/graphics/graphics.h
@@ -349,6 +349,14 @@ namespace dmGraphics
         FACE_WINDING_CW  = 1,
     };
 
+    // Face type
+    enum FaceType
+    {
+        FACE_TYPE_FRONT          = 0,
+        FACE_TYPE_BACK           = 1,
+        FACE_TYPE_FRONT_AND_BACK = 2,
+    };
+
     /*#
      * Stencil test actions.
      * Defines what happens to a stencil buffer value depending on the outcome of the stencil/depth test.
@@ -640,6 +648,69 @@ namespace dmGraphics
         TEXTURE_WRAP_REPEAT          = 3,
     };
 
+    // render states
+    enum State
+    {
+        STATE_DEPTH_TEST           = 0,
+        STATE_SCISSOR_TEST         = 1,
+        STATE_STENCIL_TEST         = 2,
+        STATE_ALPHA_TEST           = 3,
+        STATE_BLEND                = 4,
+        STATE_CULL_FACE            = 5,
+        STATE_POLYGON_OFFSET_FILL  = 6,
+        STATE_ALPHA_TEST_SUPPORTED = 7,
+    };
+
+    // buffer clear types, each value is guaranteed to be separate bits
+    enum BufferType
+    {
+        BUFFER_TYPE_COLOR0_BIT  = 0x01,
+        BUFFER_TYPE_COLOR1_BIT  = 0x02,
+        BUFFER_TYPE_COLOR2_BIT  = 0x04,
+        BUFFER_TYPE_COLOR3_BIT  = 0x08,
+        BUFFER_TYPE_DEPTH_BIT   = 0x10,
+        BUFFER_TYPE_STENCIL_BIT = 0x20,
+    };
+
+    struct PipelineState
+    {
+        uint64_t m_WriteColorMask           : 4;
+        uint64_t m_WriteDepth               : 1;
+        uint64_t m_PrimtiveType             : 3;
+        // Depth Test
+        uint64_t m_DepthTestEnabled         : 1;
+        uint64_t m_DepthTestFunc            : 3;
+        // Stencil Test
+        uint64_t m_StencilEnabled           : 1;
+
+        // Front
+        uint64_t m_StencilFrontOpFail       : 3;
+        uint64_t m_StencilFrontOpPass       : 3;
+        uint64_t m_StencilFrontOpDepthFail  : 3;
+        uint64_t m_StencilFrontTestFunc     : 3;
+
+        // Back
+        uint64_t m_StencilBackOpFail        : 3;
+        uint64_t m_StencilBackOpPass        : 3;
+        uint64_t m_StencilBackOpDepthFail   : 3;
+        uint64_t m_StencilBackTestFunc      : 3;
+
+        uint64_t m_StencilWriteMask         : 8;
+        uint64_t m_StencilCompareMask       : 8;
+        uint64_t m_StencilReference         : 8;
+        // Blending
+        uint64_t m_BlendEnabled             : 1;
+        uint64_t m_BlendSrcFactor           : 4;
+        uint64_t m_BlendDstFactor           : 4;
+        // Culling
+        uint64_t m_CullFaceEnabled          : 1;
+        uint64_t m_CullFaceType             : 2;
+        // Face winding
+        uint64_t m_FaceWinding              : 1;
+        // Polygon offset
+        uint64_t m_PolygonOffsetFillEnabled : 1;
+    };
+
     /*#
      * Texture creation parameters.
      *
@@ -756,6 +827,23 @@ namespace dmGraphics
         uint8_t  m_LayerCount; // For array texture, this is page count
         uint8_t  m_MipMap    : 7;
         uint8_t  m_SubUpdate : 1;
+    };
+
+    struct RenderTargetCreationParams
+    {
+        TextureCreationParams m_ColorBufferCreationParams[MAX_BUFFER_COLOR_ATTACHMENTS];
+        TextureCreationParams m_DepthBufferCreationParams;
+        TextureCreationParams m_StencilBufferCreationParams;
+        TextureParams         m_ColorBufferParams[MAX_BUFFER_COLOR_ATTACHMENTS];
+        TextureParams         m_DepthBufferParams;
+        TextureParams         m_StencilBufferParams;
+        AttachmentOp          m_ColorBufferLoadOps[MAX_BUFFER_COLOR_ATTACHMENTS];
+        AttachmentOp          m_ColorBufferStoreOps[MAX_BUFFER_COLOR_ATTACHMENTS];
+        float                 m_ColorBufferClearValue[MAX_BUFFER_COLOR_ATTACHMENTS][4];
+        // TODO: Depth/Stencil
+
+        uint8_t               m_DepthTexture   : 1;
+        uint8_t               m_StencilTexture : 1;
     };
 
     /*#
@@ -1225,6 +1313,73 @@ namespace dmGraphics
      * @return max_texture_size [type:uint32_t] Maximum texture size supported by GPU
      */
     uint32_t GetMaxTextureSize(HContext context);
+
+    /**
+     * Return the width of the opened window, if any.
+     * @param context Graphics context handle
+     * @return Width of the window. If no window is opened, 0 is always returned.
+     */
+    uint32_t GetWindowWidth(HContext context);
+
+    /**
+     * Return the height of the opened window, if any.
+     * @param context Graphics context handle
+     * @return Height of the window. If no window is opened, 0 is always returned.
+     */
+    uint32_t GetWindowHeight(HContext context);
+
+    /**
+     * Returns the specified width of the opened window, which might differ from the actual window width.
+     *
+     * @param context Graphics context handle
+     * @return Specified width of the window. If no window is opened, 0 is always returned.
+     */
+    uint32_t GetWidth(HContext context);
+
+    /**
+     * Returns the specified height of the opened window, which might differ from the actual window width.
+     *
+     * @param context Graphics context handle
+     * @return Specified height of the window. If no window is opened, 0 is always returned.
+     */
+    uint32_t GetHeight(HContext context);
+
+    void SetStencilMask(HContext context, uint32_t mask);
+
+    void SetStencilFunc(HContext context, CompareFunc func, uint32_t ref, uint32_t mask);
+    void SetStencilFuncSeparate(HContext context, FaceType face_type, CompareFunc func, uint32_t ref, uint32_t mask);
+    void SetStencilOp(HContext context, StencilOp sfail, StencilOp dpfail, StencilOp dppass);
+    void SetStencilOpSeparate(HContext context, FaceType face_type, StencilOp sfail, StencilOp dpfail, StencilOp dppass);
+
+    void EnableState(HContext context, State state);
+    void DisableState(HContext context, State state);
+    void SetBlendFunc(HContext context, BlendFactor source_factor, BlendFactor destinaton_factor);
+    void SetColorMask(HContext context, bool red, bool green, bool blue, bool alpha);
+    void SetDepthMask(HContext context, bool mask);
+    void SetDepthFunc(HContext context, CompareFunc func);
+
+    HContext             GetInstalledContext();
+
+    HRenderTarget NewRenderTarget(HContext context, uint32_t buffer_type_flags, const RenderTargetCreationParams params);
+    void          DeleteRenderTarget(HRenderTarget render_target);
+    void          SetRenderTarget(HContext context, HRenderTarget render_target, uint32_t transient_buffer_types);
+    HTexture      GetRenderTargetTexture(HRenderTarget render_target, BufferType buffer_type);
+    void          GetRenderTargetSize(HRenderTarget render_target, BufferType buffer_type, uint32_t& width, uint32_t& height);
+    void          SetRenderTargetSize(HRenderTarget render_target, uint32_t width, uint32_t height);
+
+    PipelineState GetPipelineState(HContext context);
+
+    void SetCullFace(HContext context, FaceType face_type);
+
+    void                 RepackRGBToRGBA(uint32_t num_pixels, uint8_t* rgb, uint8_t* rgba);
+
+    /**
+     * Get the scale factor of the display.
+     * The display scale factor is usally 1.0 but will for instance be 2.0 on a macOS Retina display.
+     * @return Scale factor
+     */
+    float GetDisplayScaleFactor(HContext context);
+
 }
 
 #endif // DMSDK_GRAPHICS_H
